+++ src/constants/StaticFolderPlaceholder.ts
@@ -1,0 +1,8 @@
+
+
+export const STATIC_FOLDER_PLACEHOLDER = {
+  hexo: {
+    postsFolder: "source/_posts",
+    placeholder: "hexo:post_asset_folder",
+  }
+}