+++ src/dashboardWebView/state/atom/AllContentFoldersAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const AllContentFoldersAtom = atom<string[] | undefined>({
+  key: 'AllContentFoldersAtom',
+  default: undefined
+});