+++ src/dashboardWebView/state/atom/AllStaticFoldersAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const AllStaticFoldersAtom = atom<string[] | undefined>({
+  key: 'AllStaticFoldersAtom',
+  default: undefined
+});