+++ src/dashboardWebView/state/atom/index.ts
@@ -1,3 +1,5 @@
+export * from './AllContentFoldersAtom';
+export * from './AllStaticFoldersAtom';
 export * from './CategoryAtom';
 export * from './DashboardViewAtom';
 export * from './FolderAtom';